Workaround Safari optimizer bug.
It appears that Safari optimizes away the arguments when they are passed to
dispatch.apply; the fix is to inline dispatch.apply in dispatch.call. I was
tempted to use slice.call(arguments, 2) here, and that works, but it also
impedes any optimization as described here:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/arguments

I can’t know for sure whether this completely avoids the Safari bug, but it at
least passes the test case here:

http://bl.ocks.org/mbostock/3177dec030e92dfd6bf08c89bdef080a

Thanks to @Hypercubed for the report.